pyproject.toml
[tool.poetry]
name = "picture_chat"
version = "0.1.0"
description = ""
authors = ["SamuelLorrain <samuel.lorrain@gmail.com>"]
readme = "README.md"
packages = [{include = "picture_chat"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
yoyo-migrations = "^8.2.0"
pyjwt = "^2.7.0"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
